refactor: 패키지 구조 변경
1. Link 클래스를 common 패키지로 이동시킨다
