+++ src/main/java/com/gaejangmo/apiserver/model/product/service/ProductService.java
@@ -1,5 +1,6 @@
 package com.gaejangmo.apiserver.model.product.service;
 
+import com.gaejangmo.apiserver.model.common.domain.vo.Link;
 import com.gaejangmo.apiserver.model.product.domain.Product;
 import com.gaejangmo.apiserver.model.product.domain.ProductRepository;
 import com.gaejangmo.apiserver.model.product.domain.vo.*;
